move secrets manager connector to its own package
